requirements.txt
airflow==1.7.1.3
alembic==0.8.10
appnope==0.1.0
Babel==1.3
bitarray==0.8.1
chartkick==0.4.2
croniter==0.3.15
decorator==4.0.11
dill==0.2.6
docutils==0.13.1
Flask==0.10.1
Flask-Admin==1.4.0
Flask-Cache==0.13.1
Flask-Login==0.2.11
Flask-WTF==0.12
funcsigs==0.4
future==0.15.2
gunicorn==19.3.0
hive-thrift-py==0.0.1
impyla==0.14.0
ipython==5.3.0
ipython-genutils==0.2.0
itsdangerous==0.24
JayDeBeApi==1.0.0
Jinja2==2.9.5
JPype1==0.6.2
lockfile==0.12.2
Mako==1.0.6
Markdown==2.6.8
MarkupSafe==0.23
numpy==1.12.0
pandas==0.19.2
pexpect==4.2.1
pickleshare==0.7.4
pluggy==0.4.0
ply==3.10
prompt-toolkit==1.0.13
protobuf==3.2.0
ptyprocess==0.5.1
py==1.4.32
Pygments==2.2.0
PyHive==0.2.1
pymongo==3.4.0
pytest==3.0.6
python-daemon==2.1.2
python-dateutil==2.6.0
python-editor==1.0.3
pytz==2016.10
PyYAML==3.12
requests==2.13.0
setproctitle==1.1.10
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
snakebite==2.11.0
SQLAlchemy==1.1.6
thrift==0.9.3
thriftpy==0.3.9
tox==2.6.0
traitlets==4.3.2
unicodecsv==0.14.1
virtualenv==15.1.0
wcwidth==0.1.7
Werkzeug==0.11.15
WTForms==2.1
